+++ security-command-center/snippets/src/main/java/v2/muteconfig/UpdateMuteRule.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package drz.muteconfig;
+package v2.muteconfig;
 
 // [START securitycenter_update_mute_config_v2]
 
@@ -31,10 +31,9 @@ public class UpdateMuteRule {
     // projectId: Google Cloud Project id.
     String projectId = "google-cloud-project-id";
 
-    // Specify the DRZ location of the mute config to update. If the mute config was
+    // Specify the location of the mute config to update. If the mute config was
     // created with v1 API, it can be accessed with "global".
-    // Available locations: "us", "eu", "global".
-    String location = "us";
+    String location = "global";
 
     // muteConfigId: Name of the mute config to update.
     String muteConfigId = "mute-config-id";
@@ -68,7 +67,7 @@ public class UpdateMuteRule {
               // If empty, all mutable fields will be updated.
               // Make sure that the mask fields match the properties changed in 'updateMuteConfig'.
               // For more info on constructing update mask path, see the proto or:
-              // https://cloud.google.com/security-command-center/docs/reference/rest/v1/folders.muteConfigs/patch?hl=en#query-parameters
+              // https://cloud.google.com/security-command-center/docs/reference/rest/v2/folders.muteConfigs/patch?hl=en#query-parameters
               .setUpdateMask(FieldMask.newBuilder().addPaths("description").build())
               .build();
